+++ pkg/storage/chunk/openstack/swift_object_client_test.go
@@ -99,8 +99,9 @@ func Test_Hedging(t *testing.T) {
 					RequestTimeout: 10 * time.Second,
 				},
 			}, hedging.Config{
-				At:   tc.hedgeAt,
-				UpTo: tc.upTo,
+				At:           tc.hedgeAt,
+				UpTo:         tc.upTo,
+				MaxPerSecond: 1000,
 			})
 			require.NoError(t, err)
 			tc.do(c)